+++ jgclark.DashboardReact/src/react/components/Switch.jsx
@@ -3,20 +3,25 @@
 
 // @flow
 import React from 'react'
-
-type SwitchProps = {
-  label: string,
-  checked: boolean,
-  onChange: (event: SyntheticInputEvent<HTMLInputElement>) => void,
-}
+import {logDebug} from '@helpers/react/reactDev.js'
 
 const Switch = ({ label, checked, onChange }: SwitchProps): React$Node => {
   return (
     <div className="switch-line">
       <label className="switch-label">{label}</label>
-      <input type="checkbox" className="apple-switch switch-input" onChange={onChange} checked={checked} />
+      <input
+        key={checked}
+        type="checkbox"
+        className="apple-switch switch-input"
+        onChange={(e) => {
+          logDebug('Switch',`${label} clicked`, e.target.checked)
+          onChange(e)
+        }}
+        checked={checked}
+      />
     </div>
   )
 }
+
 
 export default Switch